requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

evaluate==0.4.0
lavis-unofficial==0.1.10
matplotlib==3.7.0
pandas==1.5.3
Pillow==9.4.0
tqdm==4.64.1
transformers==4.24.0
sentencepiece==0.1.99
accelerate== 0.20.3